Move MySQL keyset repository into `saga` subpackage.
